Revert "Subrelease/star stories"
