--- reference-assets/Revise or Create Your Own Realms
@@ -1,87 +1,0 @@
-ADOBE INDESIGN PRINTING INSTRUCTIONS FOR SERVICE PROVIDER REPORT
-
-PUBLICATION NAME: Rolling Realms.indd
-
-PACKAGE DATE: 4/8/2020 8:45 AM
-Creation Date: 3/21/2020
-Modification Date: 4/8/2020
-
-CONTACT INFORMATION
-
-Company Name: 
-Contact: 
-Address: 
-
-
-
-
-
-Phone: 
-Fax: 
-Email: 
-
-SPECIAL INSTRUCTIONS AND OTHER NOTES
-
-
-
-
-
-
-External Plug-ins 0
-Non Opaque Objects :On PagePB, 1
-
-FONTS
-3 Fonts Used; 0 Missing, 0 Embedded, 0 Incomplete, 0 Protected
-
-Fonts Packaged
-- Name: ACaslonPro-Semibold; Type: OpenType Type 1, Status: OK
-- Name: LeagueGothic-Italic; Type: OpenType Type 1, Status: OK
-- Name: LeagueGothic-Regular; Type: OpenType Type 1, Status: OK
-
-
-COLORS AND INKS
-4 Process Inks; 0 Spot Inks
-
-- Name and Type: Process Cyan; Angle: 0.000; Lines/Inch: 0.000
-- Name and Type: Process Magenta; Angle: 0.000; Lines/Inch: 0.000
-- Name and Type: Process Yellow; Angle: 0.000; Lines/Inch: 0.000
-- Name and Type: Process Black; Angle: 0.000; Lines/Inch: 0.000
-
-
-LINKS AND IMAGES
-(Missing & Embedded Links Only)
-Links and Images: 339 Links Found; 0 Modified, 0 Missing
-Images: 0 Embedded, 339 use RGB color space
-
-
-PRINT SETTINGS
-PPD: N/A, (NPID28D75 (HP LaserJet 200 color M251nw))
-Printing To: Printer
-Number of Copies: 1
-Reader Spreads: No
-Even/Odd Pages: Both
-Pages: All
-Proof: No
-Tiling: None
-Scale: 100%, 100%
-Page Position: Upper Left
-Print Layers: Visible & Printable Layers
-Printer's Marks: None
-Bleed: 0 in, 0 in, 0 in, 0 in
-Color: Composite RGB
-Trapping Mode: None
-Send Image Data: Optimized Subsampling
-OPI/DCS Image Replacement: No
-Page Size: Letter
-Paper Dimensions: 51p0 x 66p0
-Orientation: Portrait
-Negative: No
-Flip Mode: Off
-
-
-FILE PACKAGE LIST
-
-1. Rolling Realms.indd; type: Adobe InDesign publication; size: 15012K
-2. ACaslonPro-Semibold.otf; type: Font file; size: 166K
-3. LeagueGothic-Italic.otf; type: Font file; size: 22K
-4. LeagueGothic-Regular.otf; type: Font file; size: 23K
