+++ src/components/App/App.scss
@@ -1,34 +1,3 @@
-// Breakpoint choices taken from Bootstrap site: https://getbootstrap.com/docs/4.1/layout/overview/#responsive-breakpoints
-// Two other sizes can be found there if these three aren't best
-// Extra small devices (portrait phones, less than 576px)
-// No media query for `xs` since this is the default
-
-html {
-  font-family: Arial, Helvetica, sans-serif;
-  font-size: 16px;
-}
-
-.appContainer {
-  padding: 5px;
-  display: grid;
-  gap: 10px;
-  grid-template-columns: auto;
-  grid-template-rows: auto;
-  .header {
-    grid-column: 1/2;
-    display: none;
-  }
-
-  .subheader {
-    grid-column: 1/2;
-    display: none;
-  }
-
-  .footer {
-    grid-column: 1/2;
-  }
-}
-
 button {
   background: none;
   outline: none;
@@ -36,31 +5,13 @@ button {
 }
 
 img {
-  width: 20px;
-  height: 20px;
+  // width: 30px;
+  // height: 30px;
 }
 
-// Medium devices (tablets, 768px and up)
-@media (min-width: 768px) {
-  .appContainer {
-    grid-template-columns: repeat(2, 1fr);
-    .header {
-      grid-column: 1/2;
-    }
-
-    .subheader {
-      grid-column: 1/2;
-    }
-
-    .footer {
-      grid-column: 1/2;
-    }
-  }
-}
-
-// Large devices (desktops, 992px and up)
-@media (min-width: 992px) {
-  .appContainer {
-    grid-template-columns: repeat(3, 1fr);
+@media (max-width: 480px) {
+  img {
+    // width: 15px;
+    // height: 15px;
   }
 }
