+++ src/components/BetweenTwoCastles/BetweenTwoCastles.js
@@ -16,78 +16,87 @@ const BetweenTwoCastles = () => {
         Fill squares from the bottom up; #s on top must be lower than
         those below. When you complete a column, gain a bonus.
       </p>
+      <div className="betweenTwoCastlesC">
+        <div className="castle castle1">
+          <div className="no-border" />
+          <div className="rewards no-border">
+            <img
+              src="/images/pumpkin.png"
+              alt="pumpkin"
+              className="reward"
+            />
+            <img
+              src="/images/pumpkin.png"
+              alt="pumpkin"
+              className="reward overlay"
+            />
+          </div>
+          <div className="no-border" />
+          <div className="no-border" />
+          {renderInput()}
+          <div className="rewards no-border">
+            <img
+              src="/images/heart.png"
+              alt="reward"
+              className="reward"
+            />
+            <img
+              src="/images/heart.png"
+              alt="reward"
+              className="reward overlay"
+            />
+          </div>
+          <div className="rewards no-border">
+            <img
+              src="/images/pumpkin.png"
+              alt="reward"
+              className="reward"
+            />
+          </div>
+          {renderInput()}
+          {renderInput()}
+          {renderInput()}
+          <div className="door">
+            <img src="/images/wooden-door.png" alt="door" />
+          </div>
+          {renderInput()}
+        </div>
 
-      <b />
-      <div className="rewards no-border">
-        <img
-          src="/images/pumpkin.png"
-          alt="pumpkin"
-          className="reward"
-        />
-        <img
-          src="/images/pumpkin.png"
-          alt="pumpkin"
-          className="reward overlay"
-        />
+        <div className="castle castle2">
+          <div className="rewards no-border">
+            <img
+              src="/images/coin.png"
+              alt="reward"
+              className="reward"
+            />
+            <img
+              src="/images/heart.png"
+              alt="reward"
+              className="reward overlay"
+            />
+          </div>
+          <div className="rewards no-border">
+            <img
+              src="/images/coin.png"
+              alt="reward"
+              className="reward"
+            />
+            <img
+              src="/images/coin.png"
+              alt="reward"
+              className="reward overlay"
+            />
+          </div>
+          {renderInput()}
+          {renderInput()}
+          {renderInput()}
+          {renderInput()}
+          {renderInput()}
+          <div className="door">
+            <img src="/images/stone-door.png" alt="door" />
+          </div>
+        </div>
       </div>
-      <b />
-      <b />
-      <div className="rewards no-border">
-        <img src="/images/coin.png" alt="reward" className="reward" />
-        <img
-          src="/images/heart.png"
-          alt="reward"
-          className="reward overlay"
-        />
-      </div>
-      <div className="rewards no-border">
-        <img src="/images/coin.png" alt="reward" className="reward" />
-        <img
-          src="/images/coin.png"
-          alt="reward"
-          className="reward overlay"
-        />
-      </div>
-      <b />
-      {renderInput()}
-      <div className="rewards no-border">
-        <img
-          src="/images/heart.png"
-          alt="reward"
-          className="reward"
-        />
-        <img
-          src="/images/heart.png"
-          alt="reward"
-          className="reward overlay"
-        />
-      </div>
-      <b />
-      {renderInput()}
-      {renderInput()}
-      <div className="rewards no-border">
-        <img
-          src="/images/pumpkin.png"
-          alt="reward"
-          className="reward"
-        />
-      </div>
-      {renderInput()}
-      {renderInput()}
-      <b />
-      {renderInput()}
-      {renderInput()}
-      {renderInput()}
-      <div className="door">
-        <img src="/images/wooden-door.png" alt="door" />
-      </div>
-      {renderInput()}
-      <b />
-      {renderInput()}
-      <div className="door">
-        <img src="/images/stone-door.png" alt="door" />
-      </div>
-
       <p className="margin20">
         Score 1 star per completed row (each castle has 3 rows).
       </p>
