+++ src/components/BetweenTwoCities/BetweenTwoCities.js
@@ -15,13 +15,9 @@ const BetweenTwoCities = () => {
 
   const renderCell = (i) => {
     return (
-      <input
-        key={`city-cell-${i}`}
-        type="number"
-        className=""
-        min="1"
-        max="6"
-      />
+      <div key={i} className="input">
+        <input type="number" className="" min="1" max="6" />
+      </div>
     );
   };
 
@@ -31,55 +27,47 @@ const BetweenTwoCities = () => {
         Fill a square; same #s can’t be orthogonally adjacent. Gain a
         bonus when you complete a row/column.
       </p>
-      {/* Render a 3x3 grid of inputs with two overlayed rewards appended to the end of each row */}
-      {times(9, (i) => {
-        if ((i + 1) % 3 === 1) {
-          return [<b key={`blank-${i}`} />, renderCell(i)];
-        }
-        if ((i + 1) % 3 === 2) {
-          return renderCell(i);
-        }
-        if ((i + 1) % 3 === 0) {
-          return [
-            renderCell(i),
+      <div className="gridGuideAlign">
+        <div className="betweenTwoCitiesC">
+          {times(9, (i) =>
+            (i + 1) % 3
+              ? renderCell(i)
+              : [
+                  renderCell(i),
+                  <div key={`rewards-${i}`} className="rewards">
+                    <img
+                      key={`reward-a-${i}`}
+                      className="reward"
+                      src={`/images/${rowRewards[`a${i}`]}.png`}
+                      alt={rowRewards[i - 4]}
+                    />
+                    <img
+                      key={`reward-b-${i}`}
+                      className="reward overlay"
+                      src={`/images/${rowRewards[`b${i}`]}.png`}
+                      alt={rowRewards[i - 3]}
+                    />
+                  </div>,
+                ],
+          )}
+          {times(3, (i) => (
             <div key={`rewards-${i}`} className="rewards">
               <img
                 key={`reward-a-${i}`}
                 className="reward"
-                src={`/images/${rowRewards[`a${i}`]}.png`}
-                alt={rowRewards[i - 4]}
+                src={`/images/${columnRewards[i]}.png`}
+                alt="reward"
               />
               <img
                 key={`reward-b-${i}`}
                 className="reward overlay"
-                src={`/images/${rowRewards[`b${i}`]}.png`}
-                alt={rowRewards[i - 3]}
+                src={`/images/${columnRewards[i]}.png`}
+                alt="reward"
               />
-            </div>,
-            <b key={`blank-${i}`} />,
-          ];
-        }
-        return null;
-      })}
-      <b key="blank-bottom-start" />
-      {times(3, (i) => (
-        <div key={`rewards-${i}`} className="rewards">
-          <img
-            key={`reward-a-${i}`}
-            className="reward"
-            src={`/images/${columnRewards[i]}.png`}
-            alt="reward"
-          />
-          <img
-            key={`reward-b-${i}`}
-            className="reward overlay"
-            src={`/images/${columnRewards[i]}.png`}
-            alt="reward"
-          />
+            </div>
+          ))}
         </div>
-      ))}
-      <b key="blank-bottom-end-1" />
-      <b key="blank-bottom-end-2" />
+      </div>
       <p>
         Score stars equal to the lowest of the other 2 realms this
         round. This # cannot be higher than filled squares here.
