+++ src/components/Euphoria/Euphoria.js
@@ -1,4 +1,5 @@
 import React from "react";
+import { times } from "lodash";
 import Die from "../Die/Die";
 import "./Euphoria.scss";
 
@@ -6,8 +7,8 @@ const Euphoria = () => {
   const renderDie = (i, alignment) => {
     return (
       <Die
-        key={`euphoria-${alignment}-d6-${i}`}
-        dieFace={i}
+        key={`euphoria-${alignment}-d6-${i + 1}`}
+        dieFace={i + 1}
         dieType="standard"
         initialDieState="available"
       />
@@ -21,35 +22,54 @@ const Euphoria = () => {
         <span>mark both #s</span>. Then bonus is based on the sum of
         that area.
       </p>
-
-      {renderDie(1, "left")}
-      {renderDie(2, "left")}
-      <div>1-3: </div>
-      <img src="/images/coin.png" alt="coin" className="reward" />
-      {renderDie(1, "right")}
-      {renderDie(2, "right")}
-
-      {renderDie(3, "left")}
-      {renderDie(4, "left")}
-      <div>4-10: </div>
-      <img src="/images/star.png" alt="star" className="reward" />
-      {renderDie(3, "right")}
-      {renderDie(4, "right")}
-
-      {renderDie(5, "left")}
-      {renderDie(6, "left")}
-      <div>11+: </div>
-      <div>
-        <img src="/images/heart.png" alt="heart" className="reward" />
-        /
-        <img
-          src="/images/pumpkin.png"
-          alt="pumpkin"
-          className="reward"
-        />
+      <div className="euphoriaC">
+        <div className="table">
+          {times(6, (i) => renderDie(i, "left"))}
+        </div>
+        <div className="euphoria-mid">
+          <div>
+            <span>1-3: </span>
+            <span>
+              <img
+                src="/images/coin.png"
+                alt="coin"
+                className="reward"
+              />
+            </span>
+          </div>
+          <div>
+            <span>4-10: </span>
+            <span>
+              <img
+                src="/images/star.png"
+                alt="star"
+                className="reward"
+              />
+            </span>
+          </div>
+          <div>
+            <span>11+: </span>
+            <span>
+              <img
+                src="/images/heart.png"
+                alt="heart"
+                className="reward"
+              />
+            </span>
+            <span>/</span>
+            <span>
+              <img
+                src="/images/pumpkin.png"
+                alt="pumpkin"
+                className="reward"
+              />
+            </span>
+          </div>
+        </div>
+        <div className="table">
+          {times(6, (i) => renderDie(i, "right"))}
+        </div>
       </div>
-      {renderDie(5, "right")}
-      {renderDie(6, "right")}
     </>
   );
 };
