--- src/components/Minigame/Minigame.scss
@@ -1,33 +1,0 @@
-.Minigame {
-  grid-column: 1/2;
-  border: 1px black solid;
-  border-radius: 5px;
-  height: 300px;
-  overflow-y: scroll;
-
-  // Grid
-  display: grid;
-  grid-template-columns: repeat(6, 1fr);
-  grid-template-rows: repeat(6, 1fr);
-  justify-items: center;
-  align-items: center;
-
-  p {
-    display: none;
-  }
-}
-
-.minigameHeader {
-  grid-column: 1/7;
-}
-
-// Medium devices (tablets, 768px and up)
-@media (min-width: 768px) {
-  .Minigame {
-    grid-column: span 2;
-  }
-}
-
-// Large devices (desktops, 992px and up)
-@media (min-width: 992px) {
-}
