+++ src/components/Star/Star.stories.js
@@ -1,7 +1,7 @@
 import React from "react";
 import Star from "./Star";
 import availableStar from "../../images/star-outline.png";
-import earnedStar from "../../images/star-filled.png";
+import earnedStar from "../../images/star.png";
 
 export default {
   title: "Example/Star",
@@ -9,7 +9,7 @@ export default {
   argTypes: {},
 };
 
-const Template = (args) => <Star {...args} />; // eslint-disable-line react/jsx-props-no-spreading
+const Template = () => <Star />;
 
 export const Available = Template.bind({});
 Available.args = {
