--- src/components/Viticulture/Viticulture.scss
@@ -1,3 +1,0 @@
-.wineOrder {
-  grid-column: span 2;
-}
